some code restructuring
