+++ modules/propertybasedtesting/include/inviwo/propertybasedtesting/processors/histogram.h
@@ -30,10 +30,10 @@
 #pragma once
 
 #include <inviwo/propertybasedtesting/propertybasedtestingmoduledefine.h>
+#include <inviwo/propertybasedtesting/testproperty.h>
 #include <inviwo/propertybasedtesting/algorithm/coveringarray.h>
 #include <inviwo/propertybasedtesting/algorithm/histogramtesting.h>
 #include <inviwo/propertybasedtesting/algorithm/generatingassignments.h>
-
 
 #include <inviwo/core/common/inviwo.h>
 #include <inviwo/core/processors/processor.h>
@@ -49,171 +49,6 @@
 #include <filesystem>
 
 namespace inviwo {
-
-class TestResult;
-
-class TestProperty {
-private:
-	Property* const prop;
-public:
-	TestProperty(Property* const prop)
-		: prop(prop) {
-	}
-	
-	virtual void withOptionProperties(std::function<void(OptionPropertyInt*)>) const = 0;
-	virtual std::optional<util::PropertyEffect> getPropertyEffect(
-			std::shared_ptr<TestResult>, 
-			std::shared_ptr<TestResult>) const = 0;
-
-	virtual std::ostream& ostr(std::ostream&,
-			std::shared_ptr<TestResult>) const = 0;
-	virtual std::ostream& ostr(std::ostream&,
-			std::shared_ptr<TestResult>,
-			std::shared_ptr<TestResult>) const = 0;
-
-	virtual void setToDefault() const = 0;
-	virtual void storeDefault() = 0;
-	virtual std::vector<std::shared_ptr<PropertyAssignment>> generateAssignments() = 0;
-	Property* getProperty() const {
-		return prop;
-	}
-	virtual ~TestProperty() = default;
-};
-
-template<typename T>
-class TestPropertyTyped : public TestProperty {
-	using value_type = typename T::value_type;
-	static constexpr size_t numComponents = DataFormat<value_type>::components();
-	T* const typedProperty;
-	value_type defaultValue;
-	const std::array<OptionPropertyInt*, numComponents> effectOption;
-public:
-	void withOptionProperties(std::function<void(OptionPropertyInt*)> f) const {
-		for(auto& x : effectOption)
-			f(x);
-	}
-	std::optional<util::PropertyEffect> getPropertyEffect(
-			std::shared_ptr<TestResult> newTestResult,
-			std::shared_ptr<TestResult> oldTestResult) const;
-	std::ostream& ostr(std::ostream&,
-			std::shared_ptr<TestResult>) const;
-	std::ostream& ostr(std::ostream&,
-			std::shared_ptr<TestResult>,
-			std::shared_ptr<TestResult>) const;
-
-	TestPropertyTyped(T* prop)
-		: TestProperty(prop)
-		, typedProperty(prop)
-		, defaultValue(prop->get())
-		, effectOption([prop](){
-				std::array<OptionPropertyInt*, numComponents> res;
-				for(size_t i = 0; i < numComponents; i++) {
-					const static std::vector<OptionPropertyOption<int>> options{
-							{"EQUAL",			"EQUAL",			0},
-							{"NOT_EQUAL",		"NOT_EQUAL",		1},
-							{"LESS",			"LESS",				2},
-							{"LESS_EQUAL",		"LESS_EQUAL",		3},
-							{"GREATER",			"GREATER",			4},
-							{"GREATER_EQUAL",	"GREATER_EQUAL",	5},
-							{"ANY",				"ANY",				6},
-							{"NOT_COMPARABLE",	"NOT_COMPARABLE",	7}
-						};
-					std::string identifier = prop->getIdentifier() + "_selector_" + std::to_string(i);
-					res[i] = new OptionPropertyInt(identifier, "Comparator for increasing values", options, options.size() - 1);
-				}
-				return res;
-			}()){
-		}
-	~TestPropertyTyped() = default;
-	T* getTypedProperty() const {
-		return typedProperty;
-	}
-	void setToDefault() const {
-		typedProperty->set(defaultValue);
-	}
-	const value_type& getDefaultValue() {
-		return defaultValue;
-	}
-	void storeDefault() {
-		defaultValue = typedProperty->get();
-	}
-	std::vector<std::shared_ptr<PropertyAssignment>> generateAssignments() {
-		return _generateAssignments<T>(typedProperty);
-	}
-};
-
-class TestResult {
-	private:
-		const std::vector<std::shared_ptr<TestProperty>>& defaultValues;
-		const Test test;
-		const size_t pixels;
-	public:
-		size_t getNumberOfPixels() {
-			return pixels;
-		}
-		template<typename T>
-		const typename T::value_type& getValue(const T* prop) const {
-			for(const auto& t : test)
-				if(auto p = std::dynamic_pointer_cast<PropertyAssignmentTyped<typename T::value_type>>(t);
-						p && reinterpret_cast<T*>(p->getProperty()) == prop)
-					return p->getValue();
-			for(auto def : defaultValues)
-				if(auto p = std::dynamic_pointer_cast<TestPropertyTyped<T>>(def);
-						p && p->getProperty() == prop)
-					return p->getDefaultValue();
-			assert(false);
-		}
-		TestResult(const std::vector<std::shared_ptr<TestProperty>>& defaultValues, const Test& t, size_t val)
-			  : defaultValues(defaultValues)
-			  , test(t)
-			  , pixels(val) {
-			  }
-};
-	
-template<typename T>
-std::optional<util::PropertyEffect> TestPropertyTyped<T>::getPropertyEffect(
-		std::shared_ptr<TestResult> newTestResult,
-		std::shared_ptr<TestResult> oldTestResult) const {
-	const value_type& valNew = newTestResult->getValue(this->typedProperty);
-	const value_type& valOld = oldTestResult->getValue(this->typedProperty);
-
-	std::array<util::PropertyEffect, numComponents> selectedEffects;
-	for(size_t i = 0; i < numComponents; i++)
-		selectedEffects[i] = util::PropertyEffect(effectOption[i]->getSelectedValue());
-
-	std::optional<util::PropertyEffect> res = {util::PropertyEffect::ANY};
-	for(size_t i = 0; res && i < numComponents; i++) {
-		auto compEff = util::propertyEffect(selectedEffects[i],
-				util::GetComponent<value_type, numComponents>::get(valNew, i),
-				util::GetComponent<value_type, numComponents>::get(valOld, i));
-		res = util::combine(*res, compEff);
-	}
-	return res;
-}
-
-template<typename T>
-std::ostream& TestPropertyTyped<T>::ostr(std::ostream& out,
-			std::shared_ptr<TestResult> testResult) const {
-	const value_type& val = testResult->getValue(this->typedProperty);
-	
-	std::array<util::PropertyEffect, numComponents> selectedEffects;
-	for(size_t i = 0; i < numComponents; i++)
-		selectedEffects[i] = util::PropertyEffect(effectOption[i]->getSelectedValue());
-	
-	return out << '\"' << getProperty()->getDisplayName() << "\" with identifier \"" << getProperty()->getIdentifier() << "\": "
-				<< val;
-}
-template<typename T>
-std::ostream& TestPropertyTyped<T>::ostr(std::ostream& out,
-			std::shared_ptr<TestResult> newTestResult,
-			std::shared_ptr<TestResult> oldTestResult) const {
-	const value_type& valNew = newTestResult->getValue(this->typedProperty);
-	const value_type& valOld = oldTestResult->getValue(this->typedProperty);
-	
-	return out << '\"' << getProperty()->getDisplayName() << "\" with identifier \"" << getProperty()->getIdentifier() << "\": "
-				<< valNew << ", " << valOld << " ; comparator set to  "
-				<< getPropertyEffect(newTestResult, oldTestResult);
-}
 
 /** \docpage{org.inviwo.Histogram, Histogram}
  * ![](org.inviwo.Histogram.png?classIdentifier=org.inviwo.Histogram)
@@ -261,19 +96,16 @@ private:
 	IntSizeTProperty numTests_;
 	ButtonProperty collectButton_;
 
+	void collectProperties();
 	std::vector<CompositeProperty*> compositeProperties_;
 
 	std::vector<std::shared_ptr<TestProperty>> props_;
-	void resetAllProps() {
-		for(auto prop : props_) {
-			prop->setToDefault();
-		}
-	}
+	void resetAllProps();
 
 	// Testing stuff
 	void initTesting();
 
-	bool testIsSetUp(const Test& test);
+	bool testIsSetUp(const Test& test) const;
 	void setupTest(const Test& test);
 
 	std::queue<Test> remainingTests;
