+++ modules/propertybasedtesting/src/processors/histogram.cpp
@@ -77,165 +77,156 @@ const ProcessorInfo Histogram::processorInfo_{
 };
 const ProcessorInfo Histogram::getProcessorInfo() const { return processorInfo_; }
 
-using PropertyTypes = std::tuple<OrdinalProperty<int>, IntMinMaxProperty>;
-
-struct TestablePropertyHelper {
-	template<typename T>
-	auto operator()(std::optional<std::shared_ptr<TestProperty>>& res, Property* prop) {
-		 if(auto tmp = dynamic_cast<T*>(prop); tmp != nullptr)
-			 res = {std::make_shared<TestPropertyTyped<T>>(tmp->clone())};
+void Histogram::resetAllProps() {
+	for(auto prop : props_) {
+		prop->setToDefault();
 	}
-};
-std::optional<std::shared_ptr<TestProperty>> testableProperty(Property* prop) {
-	std::optional<std::shared_ptr<TestProperty>> res = std::nullopt;
-	util::for_each_type<PropertyTypes>{}(TestablePropertyHelper{}, res, prop);
-	return res;
+}
+
+void Histogram::collectProperties() {
+	NetworkLock lock(this);
+
+	// remove previously collected properties
+	props_.clear();
+	for(auto& property : compositeProperties_)
+		removeProperty(property);
+	compositeProperties_.clear();
+
+	std::set<Property*> properties;
+
+	// collect all properties of the processors feeding directly or
+	// indirectly into this
+	for(Processor* processor : util::getPredecessors(this)) {
+		std::cerr << "visiting processor " << processor << " "
+			<< processor->getIdentifier() << std::endl;
+
+		const auto& processorProperties = processor->getProperties();
+		properties.insert(processorProperties.begin(), processorProperties.end());
+	}
+
+	// maps processor to its associated composite property
+	std::unordered_map<Processor*, CompositeProperty*> composites;
+
+	std::unordered_set<std::string> usedIdentifiers;
+	// TODO: only insert minimal set S of properties such that all properties
+	// are either in S or directly or indirectly set by a property in S
+	for(Property* prop : properties) {
+		if(auto _p = testableProperty(prop); _p != std::nullopt) {
+			props_.emplace_back(*_p);
+			auto* p = (*_p)->getProperty();
+
+			if(!usedIdentifiers.insert(p->getIdentifier()).second) {
+				size_t num = 0;
+				do {
+					p->setIdentifier(p->getIdentifier() + "_" + std::to_string(num));
+				} while(!usedIdentifiers.insert(p->getIdentifier()).second);
+			}
+
+			// add to processor
+			auto parentProcessor = dynamic_cast<Processor*>(prop->getOwner());
+			assert(parentProcessor != nullptr);
+			auto& comp = composites[parentProcessor];
+			if(comp == nullptr) {
+				std::string parentIdentifier = parentProcessor->getIdentifier();
+				std::replace(parentIdentifier.begin(), parentIdentifier.end(), ' ', '_');
+				comp = new CompositeProperty(
+						parentIdentifier,
+						parentProcessor->getDisplayName());
+				addProperty(comp);
+				compositeProperties_.emplace_back(comp);
+			}
+			comp->addProperty(p);
+			// add link to original property
+			app_->getProcessorNetwork()->addLink(p, prop);
+
+			(*_p)->withOptionProperties([&comp](auto opt){
+					comp->addProperty(opt);
+					});
+		}
+	}
 }
 
 Histogram::Histogram(InviwoApplication* app)
-    : Processor()
-    , testingState(TestingState::NONE)
-	, app_(app)
-    , tempDir_{std::filesystem::temp_directory_path() / ("inviwo_histogram_" + std::to_string(rand()))}
-	, inport_("imageInport")
-	, useDepth_("useDepth", "Use Depth", false, InvalidationLevel::Valid)
-    , color_("color", "Color", vec4(1.0f), vec4(0.0f), vec4(0.0f))
-	, countPixelsButton_("cntPixelsButton", "Count number of pixels with set color")
-	, startButton_("startButton", "Start")
-	, numTests_("numTests", "Maximum number of tests", 200, 1, 10000)
-	, collectButton_("collectButton", "Collect Properties") {
+	: Processor()
+	, testingState(TestingState::NONE)
+	  , app_(app)
+	  , tempDir_{std::filesystem::temp_directory_path() / ("inviwo_histogram_" + std::to_string(rand()))}
+	  , inport_("imageInport")
+	  , useDepth_("useDepth", "Use Depth", false, InvalidationLevel::Valid)
+	  , color_("color", "Color", vec4(1.0f), vec4(0.0f), vec4(0.0f))
+	  , countPixelsButton_("cntPixelsButton", "Count number of pixels with set color")
+	  , startButton_("startButton", "Start")
+	  , numTests_("numTests", "Maximum number of tests", 200, 1, 10000)
+	  , collectButton_("collectButton", "Collect Properties") {
 
-	countPixelsButton_.onChange([this]() {
-			app_->dispatchFrontAndForget([this]{
-					NetworkLock lock(this);
-					testingState = TestingState::SINGLE_COUNT;
-					this->invalidate(InvalidationLevel::InvalidOutput);
-			});
-		} );
+		  countPixelsButton_.onChange([this]() {
+				  app_->dispatchFrontAndForget([this]{
+						  NetworkLock lock(this);
+						  testingState = TestingState::SINGLE_COUNT;
+						  this->invalidate(InvalidationLevel::InvalidOutput);
+						  });
+				  } );
 
-	useDepth_.onChange([this]() {
-			if(useDepth_) {
-				countPixelsButton_.setDisplayName("Count number of pixels with depth value 1");
-			} else {
-				countPixelsButton_.setDisplayName("Count number of pixels with set color");
-			}
-		} );
+		  useDepth_.onChange([this]() {
+				  if(useDepth_) {
+				  countPixelsButton_.setDisplayName("Count number of pixels with depth value 1");
+				  } else {
+				  countPixelsButton_.setDisplayName("Count number of pixels with set color");
+				  }
+				  } );
 
-	startButton_.onChange([this]() {
-            initTesting();
-		} );
+		  startButton_.onChange([this]() { this->initTesting(); } );
 
-	collectButton_.onChange([this]() {
-			NetworkLock lock(this);
+		  collectButton_.onChange([this]() { this->collectProperties(); } );
 
-            // remove previously collected properties
-			props_.clear();
-            for(auto& property : compositeProperties_)
-				removeProperty(property);
-			compositeProperties_.clear();
+		  inport_.setOutportDeterminesSize(true);
+		  addPort(inport_);
 
-			std::set<Property*> properties;
+		  addProperty(useDepth_);
+		  color_.setSemantics(PropertySemantics::Color);
+		  color_.visibilityDependsOn(useDepth_, [](const auto& d) -> bool { return !d; });
+		  addProperty(color_);
 
-            // collect all properties of the processors feeding directly or
-            // indirectly into this
-			for(Processor* processor : util::getPredecessors(this)) {
-                std::cerr << "visiting processor " << processor << " "
-                          << processor->getIdentifier() << std::endl;
+		  addProperty(countPixelsButton_);
+		  addProperty(numTests_);
+		  addProperty(startButton_);
+		  addProperty(collectButton_);
 
-				const auto& processorProperties = processor->getProperties();
-				properties.insert(processorProperties.begin(), processorProperties.end());
-            }
+		  app_->dispatchFrontAndForget([this]() {
+				  // Only necessary because there is no actual serializing yet
+				  for(Property* _compProp : getProperties()) {
+				  if(auto compProp = dynamic_cast<CompositeProperty*>(_compProp); compProp != nullptr) {
+				  compositeProperties_.emplace_back(compProp);
+				  }
+				  }
+				  });
 
-			// maps processor to its associated composite property
-			std::unordered_map<Processor*, CompositeProperty*> composites;
-
-            std::unordered_set<std::string> usedIdentifiers;
-            // TODO: only insert minimal set S of properties such that all properties
-            // are either in S or directly or indirectly set by a property in S
-			for(Property* prop : properties) {
-				if(auto _p = testableProperty(prop); _p != std::nullopt) {
-					props_.emplace_back(*_p);
-					auto* p = (*_p)->getProperty();
-
-                    if(!usedIdentifiers.insert(p->getIdentifier()).second) {
-                        size_t num = 0;
-                        do {
-                            p->setIdentifier(p->getIdentifier() + "_" + std::to_string(num));
-                        } while(!usedIdentifiers.insert(p->getIdentifier()).second);
-                    }
-
-					// add to processor
-					auto parentProcessor = dynamic_cast<Processor*>(prop->getOwner());
-					assert(parentProcessor != nullptr);
-					auto& comp = composites[parentProcessor];
-					if(comp == nullptr) {
-						std::string parentIdentifier = parentProcessor->getIdentifier();
-						std::replace(parentIdentifier.begin(), parentIdentifier.end(), ' ', '_');
-						comp = new CompositeProperty(
-								parentIdentifier,
-								parentProcessor->getDisplayName());
-						addProperty(comp);
-						compositeProperties_.emplace_back(comp);
-					}
-					comp->addProperty(p);
-					// add link to original property
-					app_->getProcessorNetwork()->addLink(p, prop);
-
-					(*_p)->withOptionProperties([&comp](auto opt){
-								comp->addProperty(opt);
-							});
-				}
-			}
-		} );
-
-	inport_.setOutportDeterminesSize(true);
-	addPort(inport_);
-
-	addProperty(useDepth_);
-	color_.setSemantics(PropertySemantics::Color);
-	color_.visibilityDependsOn(useDepth_, [](const auto& d) -> bool { return !d; });
-    addProperty(color_);
-
-	addProperty(countPixelsButton_);
-	addProperty(numTests_);
-	addProperty(startButton_);
-	addProperty(collectButton_);
-
-	app_->dispatchFrontAndForget([this]() {
-            // Only necessary because there is no actual serializing yet
-            for(Property* _compProp : getProperties()) {
-                if(auto compProp = dynamic_cast<CompositeProperty*>(_compProp); compProp != nullptr) {
-                    compositeProperties_.emplace_back(compProp);
-                }
-            }
-		});
-
-	if (std::filesystem::create_directory(tempDir_)) {
-		std::stringstream str;
-		str << "Using " << tempDir_ << " to store failed tests.";
-		util::log(IVW_CONTEXT, str.str(), LogLevel::Info, LogAudience::User);
-	}
-}
+		  if (std::filesystem::create_directory(tempDir_)) {
+			  std::stringstream str;
+			  str << "Using " << tempDir_ << " to store failed tests.";
+			  util::log(IVW_CONTEXT, str.str(), LogLevel::Info, LogAudience::User);
+		  }
+	  }
 
 
 void Histogram::initTesting() {
 	testResults.clear();
 	assert(remainingTests.empty());
 
-    // store current state in order to reset it after testing
-    for(auto prop : props_) {
+	// store current state in order to reset it after testing
+	for(auto prop : props_) {
 		prop->storeDefault();
-    }
+	}
 
-    std::vector<std::shared_ptr<TestProperty>> propsToTest;
+	std::vector<std::shared_ptr<TestProperty>> propsToTest;
 	// only use those properties that have influence on this processor
-    std::copy_if(props_.begin(), props_.end(), std::back_inserter(propsToTest), [this](auto prop) {
+	std::copy_if(props_.begin(), props_.end(), std::back_inserter(propsToTest), [this](auto prop) {
 			for(auto linkedProperty : app_->getProcessorNetwork()->getPropertiesLinkedTo(prop->getProperty()))
-				if(auto parentProcessor = util::getOwningProcessor(linkedProperty))
-					if(util::getSuccessors(*parentProcessor).count(this) > 0)
-						return true;
+			if(auto parentProcessor = util::getOwningProcessor(linkedProperty))
+			if(util::getSuccessors(*parentProcessor).count(this) > 0)
+			return true;
 			return false;
-        });
+			});
 
 	std::cerr << "propsToTest.size() = " << propsToTest.size() << std::endl;
 
@@ -261,11 +252,11 @@ void Histogram::initTesting() {
 	std::cerr << "remainingTests.size() = " << remainingTests.size() << std::endl;
 	util::log(IVW_CONTEXT, std::string("Testing ") + std::to_string(remainingTests.size()) + " configurations...", LogLevel::Info, LogAudience::User);
 
-    assert(remainingTests.size() > 0);
+	assert(remainingTests.size() > 0);
 
 	app_->dispatchFront([this]() {
 			setupTest(remainingTests.front());
-		});
+			});
 }
 
 using TestingError = std::tuple<std::shared_ptr<TestResult>, std::shared_ptr<TestResult>, util::PropertyEffect, size_t, size_t>;
@@ -281,7 +272,7 @@ std::ostream& printError(std::ostream& out, const std::vector<std::shared_ptr<Te
 }
 
 void Histogram::checkTestResults() {
-    std::vector< TestingError > errors;
+	std::vector< TestingError > errors;
 
 	size_t numComparable = 0;
 	std::array<size_t,7> cnt;
@@ -355,12 +346,12 @@ void Histogram::checkTestResults() {
 			"Found " + std::to_string(numComparable) + " comparable pairs of test results",
 			LogLevel::Info, LogAudience::User);
 	if(!errors.empty()) {
-        std::stringstream str;
-        str << errors.size() << " tests failed: ";
+		std::stringstream str;
+		str << errors.size() << " tests failed: ";
 		for(size_t i = 0; i < std::min(size_t(5), errors.size()); i++) {
 			printError(str, props_, errors[i]) << std::endl;
 		}
-        util::log(IVW_CONTEXT, str.str(), LogLevel::Warn, LogAudience::User);
+		util::log(IVW_CONTEXT, str.str(), LogLevel::Warn, LogAudience::User);
 
 		// writing errors to file
 		const auto errFilePath = tempDir_ / (std::string("err_") + std::to_string(rand()) + std::string(".txt"));
@@ -369,17 +360,17 @@ void Histogram::checkTestResults() {
 			printError(errFile, props_, e) << std::endl;
 		}
 		errFile.close();
-        util::log(IVW_CONTEXT, "Wrote errors to " + errFilePath.string(), LogLevel::Info, LogAudience::User);
-    } else {
+		util::log(IVW_CONTEXT, "Wrote errors to " + errFilePath.string(), LogLevel::Info, LogAudience::User);
+	} else {
 		util::log(IVW_CONTEXT, "All tests passed.", LogLevel::Info, LogAudience::User);
 	}
 }
 
-bool Histogram::testIsSetUp(const Test& test) {
-    for(const auto& assignment : test)
+bool Histogram::testIsSetUp(const Test& test) const {
+	for(const auto& assignment : test)
 		if(!assignment->isApplied())
 			return false;
-    return true;
+	return true;
 }
 void Histogram::setupTest(const Test& test) {
 	NetworkLock lock(this);
@@ -387,22 +378,22 @@ void Histogram::setupTest(const Test& test) {
 	resetAllProps();
 
 	// then set relevant properties
-    for(const auto& assignment : test) {
+	for(const auto& assignment : test) {
 		assignment->apply();
-    }
+	}
 
 	// TODO: find better solution
 	app_->getProcessorNetwork()->forEachProcessor([](auto proc) {
 			proc->invalidate(InvalidationLevel::InvalidOutput);
-		});
+			});
 
 	app_->dispatchPool([this]() {
 			std::this_thread::sleep_for(std::chrono::milliseconds(20)); // necessary because of synchronicity issues, TODO: find better solution
 			app_->dispatchFrontAndForget([this]() {
 					testingState = TestingState::GATHERING;
 					this->invalidate(InvalidationLevel::InvalidOutput);
-				});
-		});
+					});
+			});
 }
 
 size_t countPixels(std::shared_ptr<const Image> img, const dvec4& col, const bool useDepth) {
@@ -430,15 +421,15 @@ size_t countPixels(std::shared_ptr<const Image> img, const dvec4& col, const boo
 void Histogram::process() {
 	auto img = inport_.getData();
 
-    switch(testingState) {
-        case TestingState::NONE:
+	switch(testingState) {
+		case TestingState::NONE:
 
-            break;
+			break;
 		case TestingState::SINGLE_COUNT:
-		{
-			const size_t pixelCount = countPixels(img, color_.get(), useDepth_);
-			util::log(IVW_CONTEXT, std::to_string(pixelCount) + " pixels", LogLevel::Info, LogAudience::User);
-		} break;
+			{
+				const size_t pixelCount = countPixels(img, color_.get(), useDepth_);
+				util::log(IVW_CONTEXT, std::to_string(pixelCount) + " pixels", LogLevel::Info, LogAudience::User);
+			} break;
 		case TestingState::GATHERING:
 			assert(!remainingTests.empty());
