+++ modules/propertybasedtesting/src/testproperty.cpp
@@ -1,0 +1,92 @@
+/*********************************************************************************
+ *
+ * Inviwo - Interactive Visualization Workshop
+ *
+ * Copyright (c) 2020 Inviwo Foundation
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ * list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOr
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ *
+ *********************************************************************************/
+
+#include <inviwo/propertybasedtesting/testproperty.h>
+
+namespace inviwo {
+
+template<typename T>
+std::optional<util::PropertyEffect> TestPropertyTyped<T>::getPropertyEffect(
+		std::shared_ptr<TestResult> newTestResult,
+		std::shared_ptr<TestResult> oldTestResult) const {
+	const value_type& valNew = newTestResult->getValue(this->typedProperty);
+	const value_type& valOld = oldTestResult->getValue(this->typedProperty);
+
+	std::array<util::PropertyEffect, numComponents> selectedEffects;
+	for(size_t i = 0; i < numComponents; i++)
+		selectedEffects[i] = util::PropertyEffect(effectOption[i]->getSelectedValue());
+
+	std::optional<util::PropertyEffect> res = {util::PropertyEffect::ANY};
+	for(size_t i = 0; res && i < numComponents; i++) {
+		auto compEff = util::propertyEffect(selectedEffects[i],
+				util::GetComponent<value_type, numComponents>::get(valNew, i),
+				util::GetComponent<value_type, numComponents>::get(valOld, i));
+		res = util::combine(*res, compEff);
+	}
+	return res;
+}
+
+template<typename T>
+std::ostream& TestPropertyTyped<T>::ostr(std::ostream& out,
+			std::shared_ptr<TestResult> testResult) const {
+	const value_type& val = testResult->getValue(this->typedProperty);
+	
+	std::array<util::PropertyEffect, numComponents> selectedEffects;
+	for(size_t i = 0; i < numComponents; i++)
+		selectedEffects[i] = util::PropertyEffect(effectOption[i]->getSelectedValue());
+	
+	return out << '\"' << getProperty()->getDisplayName() << "\" with identifier \"" << getProperty()->getIdentifier() << "\": "
+				<< val;
+}
+template<typename T>
+std::ostream& TestPropertyTyped<T>::ostr(std::ostream& out,
+			std::shared_ptr<TestResult> newTestResult,
+			std::shared_ptr<TestResult> oldTestResult) const {
+	const value_type& valNew = newTestResult->getValue(this->typedProperty);
+	const value_type& valOld = oldTestResult->getValue(this->typedProperty);
+	
+	return out << '\"' << getProperty()->getDisplayName() << "\" with identifier \"" << getProperty()->getIdentifier() << "\": "
+				<< valNew << ", " << valOld << " ; comparator set to  "
+				<< getPropertyEffect(newTestResult, oldTestResult);
+}
+
+struct TestablePropertyHelper {
+	template<typename T>
+	auto operator()(std::optional<std::shared_ptr<TestProperty>>& res, Property* prop) {
+		 if(auto tmp = dynamic_cast<T*>(prop); tmp != nullptr)
+			 res = {std::make_shared<TestPropertyTyped<T>>(tmp->clone())};
+	}
+};
+std::optional<std::shared_ptr<TestProperty>> testableProperty(Property* prop) {
+	std::optional<std::shared_ptr<TestProperty>> res = std::nullopt;
+	util::for_each_type<PropertyTypes>{}(TestablePropertyHelper{}, res, prop);
+	return res;
+}
+	
+} // namespace inviwo
